refactoring of cli components
